update format for spots

+++ backend/db/migrations/20230906152534-create-spot.js
@@ -42,11 +42,11 @@ module.exports = {
           allowNull: false,
         },
         lat: {
-          type: Sequelize.FLOAT,
+          type: Sequelize.DECIMAL,
           allowNull: false,
         },
         lng: {
-          type: Sequelize.FLOAT,
+          type: Sequelize.DECIMAL,
           allowNull: false,
         },
         name: {
